implemented functions

+++ include/Commodity.hh
@@ -5,6 +5,7 @@
 #include <fstream>
 #include <string>
 #include <sstream>
+#include <map>
 using namespace std;
 
 class Commodity
@@ -12,18 +13,18 @@ class Commodity
 private:
     string name;
     string type;
-    int rate;
+    double rate;
 public:
     Commodity();
-    Commodity(string, string, int);
+    Commodity(string, string, double);
     //setter methods
     void setName(string);
     void setType(string);
-    void setRate(int);
+    void setRate(double);
     //getter methods
     string getName();
     string getType();
-    int getRate();
+    double getRate();
     //other methods
     void inputDetails();
     void printDetails();
